Show go back button only on InfluencerList
